EXE-25 Implemented RMIHostService with test class

+++ executor-tester-service/src/main/java/agh/bit/ideafactory/rmi/common/TesterService.java
@@ -8,6 +8,6 @@ package agh.bit.ideafactory.rmi.common;
  */
 public interface TesterService {
 
-	public void testSubmit(int id);
+	public void testSubmit(Long id);
 
 }
